Store shuffle data in-memory via the block mgr.
As-is, shuffle data will not be deleted from the block
manager until the RDD goes out of scope. We need to add
more aggressive cleaning for this, which is necessary
now that data is stored in-memory.

This commit also increases the timeout for the Kafka
streaming tests to 10 seconds.  I'm not sure why these
are newly failing (sometimes) with the old 3 second timeout
-- but since the Kafka streaming tests are fairly unrelated
to the changes I was making, I eventually gave up on
debugging this time increase.  The streaming tests are
generally known for being flakey, so I'm not very concerned
about this.