Removed sort-based shuffle and disk spilling.
This commit removes the sort-based shuffle (which was obfuscated by
my change to move the shuffle in-memory), including
all associated unit tests.  It also removes disk spilling
that happens through the external sorter. This spilling was
used primarily as part of the sort-based shuffle, and is not
compatible with the monotasks execution model (because it
transparently spills data to disk while a task is running).  Finally, it
removes the DiskBlockObjectWriter, which is also no longer used
after all of the other removals.

As part of removing the sort-based shuffle, this change removes the
shuffle memory manager.  The shuffle memory manager was used to
ensure that not too much data was stored in memory before spilling to
disk when using sort-based shuffle.  We'll eventually need our own memory
manager, but for now, this isn't necessary, because of the removal of the
sort-based shuffle.

This change also fixes the shuffle cleanup code, which was not correct
after the last, in-memory shuffle patch.  I tried to move that change back
into the other patch, but that ended up being very difficult because it
depended on some of the code cleanup in this patch.