Disk scheduler integration.
Refactored the BlockManager to be compatible with the new DiskScheduler
and DiskMonotasks. All interaction with the disk is now done via a
DiskMonotask. In order to get a version of this commit working as
quickly as possible, several features had to be left out. The
BlockManager no longer supports dropping blocks from memory to disk in
the event that memory fills up. The BlockManager also no longer supports
removing blocks from disk or replicating blocks to other BlockManagers.
The BlockManager methods for fetching blocks from remote BlockManagers
have been left in the code, but they should be removed as soon as
possible and replaced with logic that uses NetworkMonotasks.

The CacheManager has been removed and its functionality has been moved
into the RDD class. The logic the prevents two tasks from loading the
same block at the same time has been removed.

Conflicts:
	core/src/main/scala/org/apache/spark/CacheManager.scala
	core/src/main/scala/org/apache/spark/Dependency.scala
	core/src/main/scala/org/apache/spark/broadcast/TorrentBroadcast.scala
	core/src/main/scala/org/apache/spark/monotasks/compute/ShuffleMapMonotask.scala
	core/src/main/scala/org/apache/spark/monotasks/disk/DiskReadMonotask.scala
	core/src/main/scala/org/apache/spark/monotasks/disk/DiskWriteMonotask.scala
	core/src/main/scala/org/apache/spark/monotasks/network/NetworkMonotask.scala
	core/src/main/scala/org/apache/spark/rdd/RDD.scala
	core/src/main/scala/org/apache/spark/shuffle/ShuffleReader.scala
	core/src/main/scala/org/apache/spark/storage/BlockManager.scala
	core/src/main/scala/org/apache/spark/storage/BlockManagerWorker.scala
	core/src/main/scala/org/apache/spark/storage/DiskBlockManager.scala
	core/src/main/scala/org/apache/spark/storage/DiskStore.scala
	core/src/main/scala/org/apache/spark/storage/MemoryStore.scala
	core/src/main/scala/org/apache/spark/storage/ThreadingTest.scala
	core/src/test/java/org/apache/spark/JavaAPISuite.java
	core/src/test/scala/org/apache/spark/CacheManagerSuite.scala
	core/src/test/scala/org/apache/spark/DistributedSuite.scala
	core/src/test/scala/org/apache/spark/monotasks/disk/DiskRemoveMonotaskSuite.scala
	core/src/test/scala/org/apache/spark/monotasks/disk/DiskWriteMonotaskSuite.scala
	core/src/test/scala/org/apache/spark/monotasks/network/NetworkMonotaskSuite.scala
	core/src/test/scala/org/apache/spark/rdd/PipedRDDSuite.scala
	core/src/test/scala/org/apache/spark/rdd/RDDSuite.scala
	core/src/test/scala/org/apache/spark/storage/BlockFileManagerSuite.scala
	core/src/test/scala/org/apache/spark/storage/BlockManagerSuite.scala
	streaming/src/main/scala/org/apache/spark/streaming/receiver/ReceiverSupervisorImpl.scala