Use DiskMonotasks to read HDFS files from disk.
Created a new type of DiskMonotask called HdfsReadMonotask that will
read data from HDFS and store it as serialized bytes in the
BlockManager. The NewHadoopRDD class now contains a getInputMonotasks()
method that will return an HdfsReadMonotask for each partition. The
NewHadoopRDD's compute() method is responsible for deserializing the
data that was fetched by the HdfsReadMonotask.