Disable support for org.apache.hadoop.mapred
HadoopRDD has been disabled because it does not use Monotasks to
interact with the disk, and it does not provide a separation between
disk operations and computation (deserialization, decompression). The
ability to save an RDD using the Hadoop mapred library has also been
disabled because it does not use Monotasks. NewHadoopRDD and the methods
for saving an RDD using the org.apache.hadoop.mapreduce library should
be used instead. All test cases that use the disabled functionality have
also been disabled.