Fix bugs in incrementing of computationNanos.
There were two problems with the previous tracking of nano seconds spent
on compute monotasks:

(1) Previously, the value was overwritten each time a new compute monotask was
executed, rather than incrementing the existing value.

(2) If TaskMetrics.computationNanos is incremented in
ComputeMonotask.executeAndHandleExceptions, the TaskMetrics sent back to the
driver won't reflect the nanoseconds from the last computation monotask,
because the task metrics are serialized in the last monotask's execute()
method (which is before the update to computeNanos in
ComputeMonotask.executeAndHandleExceptions).

This commit fixes both of those problems.