Added GC time to ContinuousMonitor.
The GC time can be helpful to diagnosing CPU time and other utilization,
so this commit adds it to ContinuousMonitor.

We need to be a little bit careful about this because I'm not sure what the
overhead of obtaining the GC counters are, so this may need to be removed
later on.