Add back driver heartbeater thread.
A previous commit removed the executor's heartbeat to the driver.
This heartbeat is necessary for the block manager: without this
heartbeat, the driver will eventually consider the block manager
on each machine dead. This commit adds back a minimal version of
the heartbeat that doesn't send any metrics about running
tasks.