Finished adding support for HDFS text files.
Writing text files to HDFS:

This commit creates two new monotasks: HdfsSerializationMonotask that
uses a RecordWriter to serialize an RDD and store it in the
BlockManager, and HdfsWriteMonotask that fetches a serialized RDD from
the BlockManager and writes it to HDFS. In order to tell the
RDD.buildDAG() method that an RDD needs to be written to HDFS (and that
these new monotasks need to be added to the DAG), the
PairRDDFunctions.saveAsNewAPIHadoopDataset() method sets a flag in the
RDD class that includes the necessary configuration information.

Reading text files from HDFS:

This commit fixes the logic that was first implemented in cbd6e63. After
testing that code on an EC2 instance that was running HDFS, it became
clear that the code only supported reading HDFS text files from the
local filesystem. This commit remedies this so that we now support
reading HDFS files from both HDFS and the local filesystem.

Deciding which disk queue to use for an HdfsDiskMonotask:

This commit introduces the HdfsDiskMonotask class, which is the ancestor
of all DiskMonotasks that interact with HDFS files. When interacting
with a file in the Hadoop Distributed File System, it is not possible to
determine which physical disk the file is stored on just by looking at
the file's path. This is due to the fact that HDFS file paths are an
abstraction that points to a location in the HDFS logical filesystem,
not a location on any machine's local filesystem. In order to properly
schedule HdfsDiskMonotasks, this commit adds logic to determine which
local HDFS data directory an HDFS file is stored in by looking at the
file's VolumeId. In order to determine the best disk queue to use, this
commit checks if the HDFS data directory is located on the same disk as
one of the Spark local directories, and if so, uses that Spark local
directory as the disk id.

Unfortunately, HdfsWriteMonotasks cannot know which data directory their
file will be stored in until after the file has been written. For now,
HdfsWriteMonotasks just pick a Spark local directory at random.