DiskScheduler handles DiskMonotask failures.
The DiskMonotask.execute() method now throws an exception if an error
occurs, instead of returning 'false'. Any exceptions thrown while
executing DiskMonotasks are allowed to propogate to the DiskAccessor
object handling the execution. The DiskAccessor.run() method now
contains logic to catch these exceptions and notify the
LocalDagScheduler that the DiskMonotask failed.

Also, the DiskAccessor class now handles interruptions correctly. The
main loop exits if the thread has been interrupted, and
interrupt-related exceptions are caught and handled in the proper way.

Resolves issues #24 and #25.