Created a monotask to serialize macrotask results.
Created a new ComputeMonotask called ResultSerializationMonotask that
serializes the macrotask result generated by an ExecutionMonotask. There
is exactly one ResultSerializationMonotask per macrotask, and it depends
on all of the "leaves" of the DAG of monotasks (this means that it is
always the only sink in the DAG of monotasks, and that it will be the
last monotask to run).

The reason for this change is that we need to refrain from calling
TaskMetrics.setMetricsOnTaskCompletion() until all of a macrotask's
main compute/disk/network monotasks have finished executing, otherwise
the TaskMetrics will not accurately reflect the resource usage over the
entire time that the macrotask was executing.

See issue #22.