LocalDagScheduler cleans up monotasks' temp. data.
Monotasks frequently store blocks in the BlockManager as a way of
passing temporary data to their dependents. Once all of a monotask's
dependents have finished executing, this temporary data is no longer
needed and can be deleted. This commit adds logic to the Monotasks class
that cleans up a monotask's dependencies when it is finished executing.

This required that changes be made to how the LocalDagScheduler
determines that a monotask is ready to be executed. Instead of the
LocalDagScheduler removing elements from monotasks' dependency sets,
now monotasks themselves contain the logic that determines whether their
dependencies are satisfied.

Resolves issue #8.