Instrumentation for what stage macrotasks are in.
This commit adds more metrics to the ContinuousMonitor, to record
how many macrotasks currently have at least one monotask that's
computing, performing a network transfer, or performing a disk
transfer. This helps with tracking system bottlenecks (for example,
if many macrotasks are waiting on monotasks in the network, the network
is a bottleneck).

As part of that change, this commit changes SimpleMonotask.scala (used in
testing) to subclass from ComputeMonotask rather than from Monotask, because
the LocalDagScheduler expects all monotasks to be network, compute, or disk
when determining what stage a macrotask is in. That required moving the call
to setContextClassLoader to inside ComputeMonotask.executeAndHandleExceptions,
so that SimpleMonotask could override that function to avoid needing to deal
with the class loader (there was no reason this call needed to be in the
compute scheduler and not in compute monotask in the first place).