Fix bug where free memory appears to be negative.
The BlockManagerInfo.updateBlockInfo() method assumes that a block is
only present in one storage medium (memory, disk, or tachyon) at a time.
This is not consistent with the Monotasks model, where a block that is
persisted to disk needs to be read into memory by a DiskReadMonotask
before its contents can be operated on. In this case, the block will
temporarily be stored both in memory and on disk. When the block is
eventually removed from memory, the BlockManagerInfo.updateBlockInfo()
method does not properly accounting for the new free space. As more data
is added to memory, this failure to clean up eventually causes the
BlockManagerInfo class to report negative free memory. This commit
refactors the BlockManagerInfo.updateBlockInfo() method to be consistent
with the Monotasks model.