Simplify the way blocks are fetched remotely.
In the existing code, remote shuffle blocks are fetched as follows:
(1) An open message is sent to the remote host, which returns a
unique identifier for the set of shuffle blocks to be fetched.
(2) The shuffle blocks are requested, one at a time, from the
remote host.

With monotasks, we want more fine-grained control over how shuffle
blocks are fetched, and in particular, over the precise order
in which blocks are fetched. The open message is destructive to
this goal: if first an open message is sent for A and then for
B, it's possible when the open response comes back, the thread to
process B will be scheduled first, so B's actual block(s) will be
requested before those for A.

This commit simplifies the networking code by removing the open
message. Instead, there is a single RPC to fetch a remote block:
the host that wants a remote block sents the remote host a
BlockFetchRequest, and the remote host responds with a BlockFetchSucess
message that contains the remote block (or a BlockFetchFailure message,
if the remote block could not be found).

With this change, each block corresponds to exactly one RPC to the remote
machine. This is roughly equivalent to what happened before, because with
the old networking code, one ChunkFetchRequest was sent for each block. The
main difference with this code is that each block is treated independently,
rather than treating a set of blocks that need to be fetched as a stream of
blocks that are handled together.

As part of this change, support for security has been removed (because it
adds complexity, and we don't need it right now).

In more detail, the major changes made in this pull request are:
(1) Rename BlockFetchingListener to BlockReceivedCallback.
BlockFetchingListener used to listen for multiple blocks, with a callback
each time one block was received. The new BlockReceivedCallback is called
only once (either onSuccess or onFailure) when the single block requested
is received.

(2) Remove BlockDataManager. This was previously used so Spark could use
a pluggable manager for shuffle data (so shuffle data could be fetched
either from the local block manager, or from a shuffle service); this is
no longer necessary because we’ve removed support for the shuffle service.

(3) Remove RpcHandler and all subclasses. this was previously used to handle
“OpenBlocks” message, which returned a handle to a stream of blocks that
could be fetched. We no longer fetch a stream of blocks and instead fetch
blocks one at a time, so this OpenBlocks message is no longer necessary (we
can go right to receiving blocks).

(4) Add BlockFetcher class. This is to avoid introducing a dependency
between Spark’s networking package and Spark core (The
TransportResponseHandler accepts a BlockFetcher as a parameter, which it
uses to fetch data).