Remove functionality to upload a block.
This functionality was previously used to upload blocks to remote executors.
It's no longer used because we removed support for replicating blocks remotely,
and when we re-add that functionality, it will need to work completely
differently because it will instead need to use a network monotask to transfer
the data.