Made better defaults for dev/upload.py script.
This commit changes dev/upload.py to work better by default by
making the following changes:
(1) By default, if no revision is passed in, it uses the most recent
commit.
(2) It proposes the commit message(s) from the revision used (which will be
the most recent commit by default, as described in (1)) to
use as the title / description.  The user can override the title
(the user is still prompted for a title, but there's a proposed title
that will be used if the user doesn't enter anything, similar to how
the email address currently works) but can't override the message.
I didn't add functionality to override the message because it's annoying
to do, and can easily be done in the web UI.