Change LocalDagScheduler to use an event loop.
Previously, LocalDagScheduler was thread-safe because all of the methods
that modified internal state were synchronized. This was potentially problematic
because it meant monotasks could block waiting for a LocalDagScheduler lock,
while still occupying resources. This commit changes all calls to LocalDagScheduler
to be asynchronous: external classes interact with LocalDagScheduler by posting
events to a single-threaded event loop.

Fixes issue #5.