Write all shuffle files to disk.
This commit adds functionality to write all shuffle data to disk on the map
side. This commit does not change the reduce side to read the shuffle data
from disk, which is done in a later commit. As a result, not all of the tests
pass after this commit, and running a shuffle will fail on the reduce side
(the changes are broken up in this way to facilitate easier reviewing).

A few changes needed to be made to the monotask and macrotask classes to enable
this:
-Monotask.dependencies and Monotask.dependents are now both sets, to facilitate
removal of dependencies (necessary to remove DiskWriteMonotasks that correspond
to empty shuffle blocks)
-Macrotask.getMonotasks is now implemented by subclasses of Macrotask rather than
by the Macrotask parent class, because ShuffleMapMacrotasks need to construct
the monotasks DAG very differently to include monotasks to write shuffle blocks to
disk.