Moved LocalDagSched. and DepedencyMgr to SparkEnv.
The high-level goal of this commit is to move references to globally-shared
objects that are singletons out of the TaskContextImpl and into SparkEnv.
Having references to these objects in TaskContextImpl makes it non-obvious
that these objects are singletons (instead implying that they're per-task
state); moving these objects into SparkEnv is more consistent with how other
such objects (e.g., the BlockManager) are treated in Spark. This move also makes
it much easier to run monotasks for macrotasks that are running on a remote machine
locally, because it's easier to construct a TaskContext for those monotasks (this
is necessary for the new on-disk shuffle implementation).

There are a few lower-level changes that helped enable this change:
-The LocalDagScheduler now accepts a BlockFileManager rather than a BlockManager.
The LocalDagScheduler didn't actually need the BlockManager, and only needed its
BlockFileManager, and this change means that the LocalDagScheduler and BlockManager
can both be constructed without a dependency loop (for the shuffle disk change, the
block manager needs a pointer to the LocalDagScheduler).
-Because the DependencyManager is constructed as part of making a SparkEnv, the
userClassPath parameter necessary for constructing it could be moved out of
the Executor constructor (because the executor no longer needs to construct a
dependency manager).