Removed redundant cleanup code in BlockFileManager.
The goal of this commit is to simplify how the BlockFileManager cleans
up the Spark local directories. In the past, the BlockFileManager
registered the Spark local directories to be automatically deleted
during shutdown. On top of that, it provided a public stop() method to
manually delete the Spark local directories. This method was then called
by both the BlockFileManager shutdown hook and BlockManager.stop(),
which is triggered by a call to SparkContext.stop(), which must be
called by the user. These two strategies both clear the Spark local
directories on shutdown. This commit removes the BlockFileManager's
stop() method and shutdown hook, and instead just registers the Spark
local directories to be deleted automatically.