Fixed potential race condition in BlockManager.
When getting a BlockInfo object from its internal "blockInfo" map,
the BlockManager occasionally first checks if "blockInfo" contains an
entry for the desired block, and then retrieves the entry. This two-step
process can result in a race condition if a thread removes a block's
BlockInfo object between another thread's calls to
blockInfo.contains(blockId) and blockInfo(blockId). This commit fixes
this vulnerability by always getting the BlockInfo object as an Option,
and then only processing it if it exists.

Fixed a potential race condition in
BlockManager.updateBlockInfoOnWrite() that could reorder messages to the
master.

Added comments explaining which code is responsible for synchronization.