Change resultBlockId for RddComputeMonotasks.
This commit changes the resultBlockId used by RddComputeMonotasks from
being a RDDBlockId to being a MonotaskResultBlockId. There's no reason this
result to use a RDDBlockId (because it's temporary data and not where the RDD
will more permanently be stored), and storing it with RDDBlockId can sometimes
trigger a race condition in BlockManager between when the monotask's result
gets cleaned up and when a DiskWriteMonotask writes the result
(https://github.com/NetSys/spark-monotasks/issues/26).