Encode whether to compress the block in MonotaskResultBlockId.
Usually, the block manager determines whether to compress a block
based on the type of block (e.g., one compression setting is used for
RDD blocks; another for shuffle blocks; etc.). MonotaskResultBlockIds
are used for intermediate data of all types, so this change alters
MonotaskResultBlockId to specify whether that block should be compressed.