Send reviews to dev list.
This commit changes the upload script to set monotasks-dev@groups.google.com
as the default reviewer, rather than individual people (the current members of
this group are the same as the old default reviewers). This will help us not
exceed our email quota, because all review emails will be sent to a single
email address (from the perspective of app engine) rather than to three
addresses.