Fixed issue where emailed commit messages were huge.
This issue was occuring because of the way I was passing in commit
ranges (using a colon to separate different commit message) coupled
with how the upload script parses colons. This commit fixes that issue,
so that even when a colon is used to specify a range of commits, only
the commit messages in that range are included in the code review
description.