Close and release reference to RecordReader.
This commit adds functionality to close the RecordReader and release
the reference to it in NewHadoopRdd as soon as the data has been processed.
Otherwise, we keep references to underlying buffers that store the data
that can be very large and lead to out of memory exceptions.

This commit is cherry-picked from 64135cbb3363e3b74dad3c0498cb9959c047d381
on Spark master (which fixes SPARK-9067).