Updated EC2 script.
This commit updates the EC2 script to add better defaults for
launching instances. In particular, it uses m2.4xlarge as the
default instance type, uses a newer AMI by default that has a more
recent version of Linux (note that the default AMI now only works
for pvm instance types), uses us-west-2 as the default zone (which
is usually the easiest zone to use because it's least contended),
and uses a special branch of the spark-ec2 scripts that ensures
that the disks are mounted with ext4 rather than ext3 (a consequence
of this change is that the instances take longer to mount, because
they need to be reformatted as ext4).