RDD class keeps track of BlockIds used for caching.
This commit fixes a bug where a ComputeMonotask attempts to retrieve a
cached block using its RDDBlockId when the block was actually cached by
an RddComputeMonotask using a MonotaskResultBlockId. The ComputeMonotask
is unable to find the block and therefore assumes that it needs to
recompute the RDD partition. This commit prevents this needless
recomputation by adding a map called `cachedPartitionsBlockIds` to the
RDD class, which is used to keep track of the BlockId used to cache each
partition.

This commit is necessary whenever the DAG of monotasks contains an
RdddComputeMonotask. The bug that this fixes was first discovered when
investigating why certain InputOutputMetricsSuite test cases were
failing.