Use monotasks to write Hive tables.
This commit changes the SparkHiveWriterContainer class so that it stores
a file's serialized bytes in memory. This in-memory data is then
retrieved by the InsertIntoHiveTable class and cached in the
MemoryStore. The InsertIntoHiveTable class creates an HdfsWriteMonotask
that retrieves the bytes from the MemoryStore and writes them to HDFS.

This requires three additional minor changes:
  - The MemoryStoreFileSystem class needs to be modified to support
    creating files. MemoryStoreFileSystem.create() has been refactored
    to return an FSDataOutputStream that writes to an in-memory buffer.

  - Some of the HDFS split files end up being empty. This is normal and
    also occurs in Spark. However, this requires making a small change
    to HdfsReadMonotask so that is can support reading empty files.