Simplified the design of the HDFS write logic.
This commit completely changes the design of how Monotasks saves RDDs to
HDFS. Previously, RDD.buildDag() created an HdfsSerializationMonotask
and HdfsWriteMonotask to serialize an RDD partition and write it to disk,
respectively. All that PairRDDFunctions.saveAsNewAPIHadoopDataset() had
to do was set a flag in the RDD class to notify RDD.buildDag() that the
RDD needed to be output to HDFS. This design resulted in the main job
function (which is computed by a ResultMonotask) being just a dummy
function. This, along with bugs introduced by the complex structure of
the DAG, has led us to opt for a simpler design that does not require
any changes to the DAG.

Now, the code to serialize an RDD partition is located in
PairRDDFunctions.saveAsNewAPIHadoopDataset(), and is executed in a
ResultMonotask instead of in an HdfsSerializationMonotask. This means
that the HdfsSerializationMonotask class has been removed. An
HdfsWriteMonotask is created after the data has been serialized, meaning
that it has no dependencies. In order to guarantee that the new
HdfsWriteMonotask completes before the macrotask is allowed to complete,
this commit introduces a new LocalDagScheduler event, 'AddToMacrotask',
that adds a monotask as a dependency of a macrotask's
ResultSerializationMonotask. This change requires that
LocalDagScheduler.waitingMonotasks be changed to a set of Monotask
objects, instead of a set of Monotask taskIds.

This new design allows for much of the HDFS object creation code to be
centralized in the SparkHadoopMapReduceUtil class, where in the future
it can be used by other classes.