Added support for HDFS SequenceFiles.
This commit modifies the monotasks HDFS code so that it supports writing
and reading HDFS SequenceFiles. This required making additions to
SparkHadoopMapReduceUtil.getRecordWriter() in order to create a
SequenceFile.Writer object that writes to a byte stream instead of to a
file.

Also, this commit refactors SparkContext.sequenceFile() to use the new
Monotasks HDFS API.