Further throttle assignment of tasks to workers.
Previously (as of commit 02715eac0a0c07811b74b488d9658f26a2230634), we assigned up
to (# cores + # disks + 1) tasks to each executor. The intuition behind this approach
was that the executor can't run any more tasks than that concurrently, so that is a
maximum on how many tasks are needed to keep the pipeline full.  The problem with that
number was that it was sometimes still too many tasks, meaning that tasks would queue
on the executor, and it might have been better to let the tasks queue at the scheduler,
so they could be assigned to a different executor where they could run sooner.
In practice, this was leading to significant load-imbalance in how tasks were assigned
when running experiments.

This commit adds functionality to the master scheduler to assign fewer tasks
to each executor if the tasks don't use disk and network. If tasks don't use
network, only (# cores + # disks) are assigned to each executor, if tasks don't
use disk, only (# cores + 1) are assigned to each executor, and if tasks don't
use network or disk, only (# cores) tasks are assigned. This functionality assumes
that only one task set is being scheduled at a given point in time (otherwise
this reasoning doesn't make sense).  We'll need to fix this in the future, when
we start looking at workloads where multiple jobs run concurrently.

To implement this functionality, I added metadata tracking in the RDD class to track
whether RDDs are going to be saved to HDFS, which is necessary so the scheduler can
determine whether the tasks to compute the RDD will use disk.