Write all of a task's shuffle data to 1 file.
This commit changes monotasks' shuffle implementation to
work the same way as Spark's sort-based shuffle: for each map task,
instead of writing a separate shuffle file for each reduce task, the task
writes a single file that contains the shuffle data for all of the reducers.
Data is written in order by reduce ID, and we write a separate index file
that describes the offset at which data for each reduce task is located.

There are a few issues with this change. One is that we now write all
shuffle blocks to the memory store, even when they're empty, which introduces
a significant performance degredation when running a shuffle that has a
large number of empty shuffle blocks.  The other potential improvement is that
there's no particular reason that the index file needs to be separate from
the data file: we could just write the indices at the beginning of the file where
the data is located. It might also make sense to just store the indices along
with the block manager's metadata, in memory.