Determine which disk an HdfsWriteMonotask will access.
This commit changes how HdfsWriteMonotasks are submitted so that it is
possible to determine which local disk they will write to. When an
HdfsWriteMonotask is submitted to the DiskScheduler, it is not
immediately assigned to a disk queue, but is instead placed in a
separate queue. This queue is serviced by a new thread, which
initializes each HdfsWriteMonotask. The initialization process consists
of creating the HDFS file that the monotask will write to and writing
the first byte of data, which enables its metadata, including which
local filesystem disk it is stored on, to be queried via the HDFS API.
The thread then submits the HdfsWriteMonotask to the appropriate disk
queue.