Add DiskScheduler support for multiple threads per disk
This commit introduces a configuration parameter
"spark.monotasks.threadsPerDisk" that allows the user to specify the
maximum number of threads that will be allowed to access each phyiscal
disk at the same time. This adapts the DiskScheduler to better support
Solid State Drives, which, unlike Hard Disk Drives, achieve their best
throughput when saturated with concurrent accesses.
"spark.monotasks.threadsPerDisk" defaults to 1 when unset, which is the
optimal configuration when using HDDs.