Fixed broken shuffle file cleanup.
Previously, when we cleaned up shuffle data, we only deleted the file
for ShuffleBlockId(shuffleId, mapId, 0), which resulted in all of the shuffle
data getting deleted from disk. However, the BlockManager still stored
metadata about the data with ShuffleBlockId(shuffleId, mapId, reduceId).
This commit adds functionality to delete that metadata.

In general, the treatment of shuffle data remains messy, because the logic about
mapping the BlockId for the file where the shuffle data is stored to the BlockIds
for all the blocks stored in that file is sprinkled across the code base. This
should be refactored at some point.