Remove unnecessary copy in HdfsWriteMonotask.
The HdfsWriteMonotask code previously used slice() to take a subset
of the array to be written to disk, which makes a copy of the array.
This copy can take a long time, resulting in idle time on the disk.
This commit changes it to instead pass in an offset and length when
writing the array, so that we can avoid the array copy.