Eliminated code to try to join DiskAccessors.
This code led to hangs when running on a cluster, so this commit
eliminates it. This may result in monotasks getting stopped while
a disk thread is in the middle of a write, but that is preferable to
hanging during shutdown.

Ideally we should fix this in the future so that DiskMonotasks are
stopped correctly.