HdfsWriteMonotask specifies correct write length.
Previously, HdfsWriteMonotask used the
FSDataOutputStream.write(Array[Byte], Int, Int) method incorrectly: the
code treated the third parameter as the end index instead of the length,
which meant that the call was trying to write one byte more than the
size of the array. This error only appeared occassionally because
sometimes the backing array is larger than the buffer's contents, so
trying to access the extra byte did not cause an
ArrayIndexOutOfBoundsException. In that case, the last byte to be
written was just garbage. The error only manifested when the backing
array was exactly the same size as the buffer.

The fix is simple: just use the number of bytes remaining as the third
parameter instead of the size of the buffer.