Write shuffle data into a DirectByteBuffer.
This commit changes the behavior of the shuffle such that the
ComputeMonotask that generates the serialized shuffle data writes all of the
data into a single DirectByteBuffer, rather than into one (non-direct) ByteBuffer
for each shuffe block. We also write an index at the beginning of the
buffer, which can be used to determine where individual shuffle blocks lie
within the file.

This change was necessary to eliminate CPU time that was occuring as part of
DiskWriteMonotasks. It took hundreds of milliseconds (for ~100MB blocks) just to
copy the data to be written from multiple buffers in the JVM to an off-heap buffer
that could be written to disk, which meant that DiskWriteMonotasks weren't fully
utilizing the disk (there were gaps at the beginning of each DiskWriteMonotask,
where the disk wasn't used at all). By eliminating some of the CPU time that occured
as part of DiskWriteMonotasks, this commit reduces the time to run a disk-intensive
map stage by about 2% (I verified this in two different expermients with
different amounts of data, and in each experiment I took the average over 10 trials).

Unfortunately this commit does not completely eliminate the CPU time; still, roughly
$5$\% of the time in the disk monotasks is still spent copying data. We could potentially
eliminate most of the remaining overhead by memory mapping the output file and writing
the data there, but that involves some complexities in making sure the buffer cache
doesn't intervene and start the disk write early.

This change also uses a new type of BlockId, MultipleShuffleBlocksId, to write a block
that consists of multiple shuffle blocks. We wrote blocks with multiple shuffle blocks
in the old code and used a ShuffleBlockId with reduce ID 0, which was confusing,
because it is the same block ID that is used for the data that is for only reduce ID 0.
Using an explicit BlockId makes these two types of blocks easier to differentiate
(both in coding and in debugging).

This commit doesn't implement one thing that will be important to do eventually: it
doesn't make any attempt to try to keep track of the memory that we allocate off-heap.
We should definitely manage this somehow, so that we can avoid using too much or at
least cleanly exit with an error when we exceed the amount of off-heap memory available.